+++ src/pptx2/shapes/shapetree.py
@@ -695,7 +695,7 @@ class _BaseGroupShapes(_BaseShapes):
 
     def add_picture(
         self,
-        image_file: str | IO[bytes],
+        image_file: str | os.PathLike[str] | IO[bytes],
         left: Length = Emu(0),
         top: Length = Emu(0),
         width: Length | None = None,
@@ -707,8 +707,9 @@ class _BaseGroupShapes(_BaseShapes):
     ) -> Picture:
         """Add picture shape displaying image in `image_file`.
 
-        `image_file` can be either a path to a file (a string) or a file-like object. The picture
-        is positioned with its top-left corner at (`left`, `top`). If `width` and `height` are
+        `image_file` can be either a path to a file (a string or `pathlib.Path`) or a
+        file-like object. The picture is positioned with its top-left corner at
+        (`left`, `top`). If `width` and `height` are
         both |None|, the native size of the image is used. If only one of `width` or `height` is
         used, the unspecified dimension is calculated to preserve the aspect ratio of the image.
         If both are specified, the picture is stretched to fit, without regard to its native